+++ src/main/java/lotto/views/Input.java
@@ -2,7 +2,6 @@ package lotto.views;
 
 import camp.nextstep.edu.missionutils.Console;
 import lotto.utils.Constant;
-import lotto.validation.InputValidation;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -10,11 +9,9 @@ import java.util.List;
 
 public class Input {
 
-    private final InputValidation inputValidation;
     private final Output output;
 
     public Input() {
-        inputValidation = new InputValidation();
         output = new Output();
     }
 
@@ -26,33 +23,12 @@ public class Input {
         return question;
     }
 
-    public List<Integer> getWinningNumber() {
+    public String getWinningNumber() {
         output.printGetWinningNumbers();
 
         String answer = Console.readLine();
-        inputValidation.checkNumber(answer, Constant.REGEX_WINNING_NUMBER_INPUT);
-        List<Integer> result = convertToList(answer);
-        inputValidation.checkNumberSize(result);
-        inputValidation.checkDuplicate(result);
-        inputValidation.checkNumberInRange(result);
-        sortWinningNumber(result);
 
-        return result;
-    }
-
-    public List<Integer> convertToList(String numbers) {
-        String[] split = numbers.split(Constant.REGEX_WINNING_NUMBER_INPUT);
-        List<Integer> result = new ArrayList<>();
-
-        for (int i = 0; i < split.length; i++) {
-            result.add(Integer.parseInt(split[i]));
-        }
-
-        return result;
-    }
-
-    public void sortWinningNumber(List<Integer> numbers) {
-        Collections.sort(numbers);
+        return answer;
     }
 
     public String getBonusNumber() {
