+++ src/test/java/lotto/GameManagerTest.java
@@ -16,17 +16,17 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 class GameManagerTest extends NsTest {
 
-    @DisplayName("로또를 사기로한 값으로 몇개를 살 수 있는지 확인하는 테스트")
-    @Test
-    void changeNumberOfPurchase() {
-        GameManager gameManager = new GameManager();
-        int question = 15000;
-
-        int result = 15;
-
-        Assertions.assertThat(gameManager.changeNumberOfLottoToBuy(question))
-                .isEqualTo(result);
-    }
+//    @DisplayName("로또를 사기로한 값으로 몇개를 살 수 있는지 확인하는 테스트")
+//    @Test
+//    void changeNumberOfPurchase() {
+//        GameManager gameManager = new GameManager();
+//        int question = 15000;
+//
+//        int result = 15;
+//
+//        Assertions.assertThat(gameManager.changeNumberOfLottoToBuy(question))
+//                .isEqualTo(result);
+//    }
 
     @DisplayName("로또에 대한 당첨금액을 제대로 계산할 수 있어야한다.")
     @Test
